+++ framework/model/src/main/java/is/codion/framework/model/DefaultEntitySearchModel.java
@@ -48,7 +48,7 @@ import java.util.function.Supplier;
 
 import static is.codion.framework.domain.entity.condition.Condition.and;
 import static is.codion.framework.domain.entity.condition.Condition.or;
-import static is.codion.framework.model.EntityEditModel.editEvents;
+import static is.codion.framework.model.EntityEditModel.events;
 import static java.text.MessageFormat.format;
 import static java.util.Collections.*;
 import static java.util.Objects.requireNonNull;
@@ -92,8 +92,8 @@ final class DefaultEntitySearchModel implements EntitySearchModel {
 						.collect(toMap(Function.identity(), column -> new DefaultSettings())));
 		this.limit = Value.nullable(builder.limit);
 		if (builder.editEvents) {
-			editEvents().updated(entityDefinition.type()).addWeakConsumer(updateListener);
-			editEvents().deleted(entityDefinition.type()).addWeakConsumer(deleteListener);
+			events().updated(entityDefinition.type()).addWeakConsumer(updateListener);
+			events().deleted(entityDefinition.type()).addWeakConsumer(deleteListener);
 		}
 	}